+++ base/sparse/sparse.jl
@@ -31,7 +31,7 @@ import Base: @get!, acos, acosd, acot, acotd, acsch, asech, asin, asind, asinh,
 
 export AbstractSparseArray, AbstractSparseMatrix, AbstractSparseVector,
     SparseMatrixCSC, SparseVector, blkdiag, droptol!, dropzeros!, dropzeros,
-    issparse, nonzeros, nzrange, rowvals, sparse, sparsevec, spdiagm, speye, spones,
+    issparse, nonzeros, nzrange, rowvals, sparse, sparsevec, spdiagm, spones,
     sprand, sprandn, spzeros, nnz, permute
 
 include("abstractsparse.jl")